requirements.txt
Django==3.1.14
djangorestframework==3.13.1
six==1.16.0
river==0.11.1
azure.storage.blob==12.13.1
pandas==1.4.3 
matplotlib==3.5.3
django-storages-azure==1.6.8
azure-data-tables==12.4.0
scikit-learn==1.1.2
pyyaml==6.0
asgiref==3.5.2
azure-common==1.1.28
azure-core==1.25.0
azure-storage-common==2.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
idna==3.3
isodate==0.6.1
kiwisolver==1.4.4
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pip==22.0.4
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.0
setuptools==58.1.0
sqlparse==0.4.2
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.11